Update apt-cacher-ng config and setup scripts
This change makes it safer to use apt-cacher-ng and pbuilder over hostile
networks. apt-cacher-ng now redirects http requests to https for both debian
and subgraph repositories. Update the setup and cow-setup scripts to use these
changes.

This change was needed because of some wildly unsafe behavior by pbuilder:
During the bootstrapping (git-pbuilder create?) process, wget is run as root
(!) for a few http urls needed for bootstrapping (Release, etc).

With these changes, wget only talks to apt-cacher-ng and apt-cacher-ng does the
actual http/https requests/processing. Effectively this ensures that the
bootstrapping process no longer running a tls library and http client as root
(!).

This is a bandaid. Upstream, the bootstrapping process should not run wget as
root for http connections to random (!) debian mirrors.